Requiring C++14 now.
More simplifications for C++14, CMake setup,
silenced some warning.

+++ Emulator/Sound/TPulseAudioSoundManager.h
@@ -99,7 +99,7 @@ private:
 
     static void SPAStreamWriteCallback(pa_stream* s, size_t requested_bytes, void* userData)
     {
-        return ((TPulseAudioSoundManager*) userData)->PAStreamWriteCallback(s, requested_bytes);
+        return ((TPulseAudioSoundManager*) userData)->PAStreamWriteCallback(s, (unsigned)requested_bytes);
     }
 
     void PAContextStateCallback(pa_context* context, pa_threaded_mainloop* mainloop);
